+++ consensus/ethash/sealer.go
@@ -258,7 +258,7 @@ type sealWork struct {
 
 func startRemoteSealer(ethash *Ethash, urls []string, noverify bool) *remoteSealer {
 	if ModePandora == ethash.config.PowMode {
-		return StartRemotePandora(ethash, urls, noverify, true)
+		return nil
 	}
 
 	ctx, cancel := context.WithCancel(context.Background())
